requirements.txt
streamlit
sentence-transformers
numpy
faiss-cpu
protobuf==3.20.3
